...could it be??? that AI was mostly a lot easier than I thought it would be???
well that was unexpected! and a welcome relief!
anywho, AI is mostly done? maybe? if I'm not an idiot? we'll have to wait and see, and make sure i'm not actually ***s t u p i d*** and messed something major up. also, a funny bit of code was found in WaterLeapAI. 10/10.

onto various other bits, such as finishing Dinosaurs' declaring classes.